Updates service to return any available data types
Now when you query for historical data, you use a path of
/historical/<data type>. This allows us to query for something other than
just altitude.

As part of this task I've refactored the code to decouple some of the
different parts. The server no longer worries about the database
implementation, that is handled by the telemetryDb module. The business
logic to get telemetry keys and to format the returned data correctly has
now been moved into the new queryHelper module.